refactor(tests): run README doctests inside Pytest
Previously, the tests used the separate `byexample` tool in order to extract and
run the shell snippets in the README file. That worked very well, but was
relatively slow.

Now, we extract and update the same tests via a custom regex, with all the
diffing and updating conveniently done by Inline-Snapshot. This works pretty
well, though it only supports the subset of features that we currently need.