fix: allow packages without versions in `uv.lock` files
Fixes https://github.com/latk/ganzua/issues/4

When a project uses Setuptools and dynamic versions and editable installs, uv
will fail to properly record the package's version in the lockfile. While this
scenario should be rare in modern projects that largely avoid Setuptools (and
avoid dynamic versions), it's an entirely legitimate use case. Since uv doesn't
seem to guarantee the presence of version information, we shouldn't assume this
either.

However, it would be exceedingly tricky to properly deal with version-less
packages. Therefore, we substitute the fake version `0+undefined`. This is a
valid version specifier, consisting of version 0 plus a local version tag.