docs: simplify example commands
* suggest `uvx` instead of `uv tool run`
* suggest `git show` instead of `git cat-file blob`

In order to set a link to the `git show` docs, mild adjustments to the CLI help
rendering where necessary. By default, Rich would try to render the Hyperlink as
a terminal link, which is not universally supported and hides the URL. Disabling
hyperlinks actually results in a nicer `text (href)` format.