feat: infer lockfiles for `ganzua inspect` and `ganzua diff`
As in the `ganzua constraint ... --lockfile` cases, the paths to lockfiles can
now just point  to the directories, with the concrete lockfile getting inferred.

For `ganzua inspect`, the lockfile argument has become optional, defaulting to
the current working directory `.`.

For `ganzua diff OLD NEW`,  both arguments are still required to avoid
confusion.