feat: track information about extras
Keep the extra names from the `[project.optional-dependencies]` and
`[tool.poetry.extras]` tables.

The `Requirement` data model gains a new `in_extras` field to store this info.
This cannot be called `extras` because those describe the extras enabled for a
package, not the extras that this Requirement is part of. This means the
`groups` and `in_extras` fields have a similar purpose, but are named
inconsistently in order to maintain compatibility.