start working on a website for better docs
Use mdbook (popular in the Rust ecosystem) because it's low-magic and supports
CommonMark.

Use the Lychee link checker to ensure that we don't have any broken links, which
I also use on my website.

Some changes to the `ganzua help` system were necessary in order to create
Markdown snippets with a structure suitable for inclusion in the docs.