show JSON Schema in docs
Instead of dumping the raw JSON, we do a bit of pattern-matching to generate
suitable Markdown from the JSON Schema file. This works surprisingly well, even
if it doesn't have full fidelity.

There are a couple of existing tools for such conversions:

* <https://github.com/adobe/jsonschema2md>
* <https://github.com/sbrunner/jsonschema2md>
* <https://github.com/elisiariocouto/jsonschema-markdown>

Many of these tools generate tables, which is unsuitable for longer docstrings.
Stéphane Brunner's tool is almost a perfect fit, but I have minor style
disagrements, and it's not worth pulling in a dependency for ~200 lines of code,
especially since we only have to support the JSON Schema flavor generated by
Pydantic.